+++ spec/Pim/Bundle/CatalogBundle/AttributeType/TextTypeSpec.php
@@ -20,7 +20,7 @@ class TextTypeSpec extends ObjectBehavior
 
     function it_builds_the_attribute_forms(FormFactory $factory, $name)
     {
-        $this->buildAttributeFormTypes($factory, $name)->shouldHaveCount(8);
+        $this->buildAttributeFormTypes($factory, $name)->shouldHaveCount(7);
     }
 
     function it_prepares_the_product_value_form($value, $name)