+++ CHANGELOG.md
@@ -24,8 +24,13 @@
 # 1.0.x
 
 ## Bug fixes
+
+# 1.0.3 (2014-10-24)
+
+## Bug fixes
 - PIM-3206: Removing the group "ALL" from categories' permissions after a clean installation.
 - PIM-3234: Fix performance issue on granted category filter
+- PIM-3300: Fixed bug on revert of a multiselect attribute options
 
 # 1.0.2 (2014-10-10)